pass 10th stage

+++ src/main/java/bittorrent/torrent/TorrentInfo.java
@@ -12,7 +12,7 @@ public record TorrentInfo(
 	byte[] hash,
 	long length,
 	String name,
-	long pieceLength,
+	int pieceLength,
 	List<byte[]> pieces
 ) {
 
@@ -20,7 +20,7 @@ public record TorrentInfo(
 		final var hash = DigestUtils.shaInfo(root);
 		final var length = (long) root.getOrDefault("length", -1l);
 		final var name = (String) root.get("name");
-		final var pieceLength = (long) root.get("piece length");
+		final var pieceLength = (int) (long) root.get("piece length");
 
 		final var pieceHashes = ((String) root.get("pieces")).getBytes(StandardCharsets.ISO_8859_1);
 		final var pieces = new ArrayList<byte[]>();
